Refact functions in engine in GO
