+++ internal/engine/elasticsearch/get.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"fmt"
 
+	"ragflow/internal/common"
 	"ragflow/internal/engine/types"
 )
 
@@ -46,4 +47,28 @@ func (e *elasticsearchEngine) GetChunk(ctx context.Context, indexName, chunkID s
 	}
 
 	return searchResp.Chunks[0], nil
+}
+
+// GetFields is not implemented for Elasticsearch
+func (e *elasticsearchEngine) GetFields(chunks []map[string]interface{}, fields []string) map[string]map[string]interface{} {
+	common.Warn("GetFields not implemented for Elasticsearch")
+	return nil
+}
+
+// GetAggregation is not implemented for Elasticsearch
+func (e *elasticsearchEngine) GetAggregation(chunks []map[string]interface{}, fieldName string) []map[string]interface{} {
+	common.Warn("GetAggregation not implemented for Elasticsearch")
+	return nil
+}
+
+// GetHighlight is not implemented for Elasticsearch
+func (e *elasticsearchEngine) GetHighlight(chunks []map[string]interface{}, keywords []string, fieldName string) map[string]string {
+	common.Warn("GetHighlight not implemented for Elasticsearch")
+	return nil
+}
+
+// GetDocIDs is not implemented for Elasticsearch
+func (e *elasticsearchEngine) GetDocIDs(chunks []map[string]interface{}) []string {
+	common.Warn("GetDocIDs not implemented for Elasticsearch")
+	return nil
 }
