+++ internal/engine/infinity/common.go
@@ -18,7 +18,6 @@ package infinity
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
 	"fmt"
 	"ragflow/internal/common"
@@ -26,103 +25,6 @@ import (
 
 	infinity "github.com/infiniflow/infinity-go-sdk"
 )
-
-// Delete deletes rows from either a dataset table or metadata table.
-// If indexName starts with "ragflow_doc_meta_", it's a metadata table.
-// Otherwise, it's a dataset table: {indexName}_{datasetID}
-func (e *infinityEngine) Delete(ctx context.Context, condition map[string]interface{}, indexName string, datasetID string) (int64, error) {
-	var tableName string
-	if strings.HasPrefix(indexName, "ragflow_doc_meta_") {
-		tableName = indexName
-	} else {
-		tableName = fmt.Sprintf("%s_%s", indexName, datasetID)
-	}
-
-	db, err := e.client.conn.GetDatabase(e.client.dbName)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get database: %w", err)
-	}
-
-	table, err := db.GetTable(tableName)
-	if err != nil {
-		common.Warn(fmt.Sprintf("Table %s does not exist, skipping delete", tableName))
-		return 0, nil
-	}
-
-	// Get table columns for building filter
-	clmns := make(map[string]struct {
-		Type    string
-		Default interface{}
-	})
-	colsResp, err := table.ShowColumns()
-	if err != nil {
-		return 0, fmt.Errorf("failed to get columns: %w", err)
-	}
-	result, ok := colsResp.(*infinity.QueryResult)
-	if ok {
-		if nameArr, ok := result.Data["name"]; ok {
-			if typeArr, ok := result.Data["type"]; ok {
-				if defArr, ok := result.Data["default"]; ok {
-					for i := 0; i < len(nameArr); i++ {
-						colName, _ := nameArr[i].(string)
-						colType, _ := typeArr[i].(string)
-						var colDefault interface{}
-						if i < len(defArr) {
-							colDefault = defArr[i]
-						}
-						clmns[colName] = struct {
-							Type    string
-							Default interface{}
-						}{colType, colDefault}
-					}
-				}
-			}
-		}
-	}
-
-	// Build filter from condition
-	filter := buildFilterFromCondition(condition, clmns)
-
-	delResp, err := table.Delete(filter)
-	if err != nil {
-		return 0, fmt.Errorf("failed to delete: %w", err)
-	}
-
-	return delResp.DeletedRows, nil
-}
-
-// DropTable deletes a table/index
-func (e *infinityEngine) DropTable(ctx context.Context, indexName string) error {
-	db, err := e.client.conn.GetDatabase(e.client.dbName)
-	if err != nil {
-		return fmt.Errorf("Failed to get database: %w", err)
-	}
-
-	_, err = db.DropTable(indexName, infinity.ConflictTypeIgnore)
-	if err != nil {
-		return fmt.Errorf("Failed to drop table: %w", err)
-	}
-	return nil
-}
-
-// TableExists checks if table/index exists
-func (e *infinityEngine) TableExists(ctx context.Context, indexName string) (bool, error) {
-	db, err := e.client.conn.GetDatabase(e.client.dbName)
-	if err != nil {
-		return false, fmt.Errorf("Failed to get database: %w", err)
-	}
-
-	_, err = db.GetTable(indexName)
-	if err != nil {
-		// Check if error is "table not found"
-		errLower := strings.ToLower(err.Error())
-		if strings.Contains(errLower, "not found") || strings.Contains(errLower, "notexist") || strings.Contains(errLower, "doesn't exist") {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
 
 // fieldInfo represents a field in the infinity mapping schema
 type fieldInfo struct {
