+++ internal/service/tenant.go
@@ -267,11 +267,9 @@ func (s *TenantService) GetTenantList(userID string) ([]*TenantListItem, error) 
 
 // CreateMetadataInDocEngine creates the document metadata table for a tenant
 func (s *TenantService) CreateMetadataInDocEngine(tenantID string) (common.ErrorCode, error) {
-	// Build table name: ragflow_doc_meta_<tenant_id>
-	tableName := fmt.Sprintf("ragflow_doc_meta_%s", tenantID)
-
 	// Call document engine to create doc meta table
-	err := s.docEngine.CreateMetadata(context.Background(), tableName)
+	// Table name will be built as "ragflow_doc_meta_{tenantID}" in the engine
+	err := s.docEngine.CreateMetadataStore(context.Background(), tenantID)
 	if err != nil {
 		return common.CodeServerError, fmt.Errorf("failed to create metadata table: %w", err)
 	}
@@ -285,7 +283,7 @@ func (s *TenantService) DeleteMetadataInDocEngine(tenantID string) (common.Error
 	tableName := fmt.Sprintf("ragflow_doc_meta_%s", tenantID)
 
 	// Call document engine to delete doc meta table
-	err := s.docEngine.DropTable(context.Background(), tableName)
+	err := s.docEngine.DropStore(context.Background(), tableName)
 	if err != nil {
 		return common.CodeServerError, fmt.Errorf("failed to delete doc meta table: %w", err)
 	}
